refresh: Check for existence of install script
This is a follow up to the patch "refresh: Check for existence of install
script bevore attemting to move it"

+++ shellpack_src/src/refresh.sh
@@ -18,7 +18,7 @@ if [ -e $1/$1-bench ]; then
 	cp $1/$1-bench ../../shellpacks/shellpack-bench-$1
 fi
 if [ -e $1/$1-install ]; then
-    cp $1/$1-install ../../shellpacks/shellpack-install-$1
+	cp $1/$1-install ../../shellpacks/shellpack-install-$1
 fi
 
 for STAGES in 1 2; do
@@ -27,9 +27,11 @@ for STAGES in 1 2; do
 		mv ../../shellpacks/shellpack-bench-$1.stage ../../shellpacks/shellpack-bench-$1
 		chmod a+x ../../shellpacks/shellpack-bench-$1
 	fi
-	cat ../../shellpacks/shellpack-install-$1 | ../bin/rewrite-shellpack $1 > ../../shellpacks/shellpack-install-$1.stage
-	mv ../../shellpacks/shellpack-install-$1.stage ../../shellpacks/shellpack-install-$1
-	chmod a+x ../../shellpacks/shellpack-install-$1
+	if [ -e ../../shellpacks/shellpack-install-$1 ]; then
+		cat ../../shellpacks/shellpack-install-$1 | ../bin/rewrite-shellpack $1 > ../../shellpacks/shellpack-install-$1.stage
+		mv ../../shellpacks/shellpack-install-$1.stage ../../shellpacks/shellpack-install-$1
+		chmod a+x ../../shellpacks/shellpack-install-$1
+	fi
 done
 
 exit 0
